rappa-ai-mvp/backend/requirements.txt
# Web Framework
fastapi==0.108.0
uvicorn[standard]==0.25.0
python-multipart==0.0.6

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Task Queue
celery==5.3.4
redis==5.0.1

# OCR & Image Processing
pytesseract==0.3.10
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.26.2

# Storage
boto3==1.34.10

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Data Processing & Export
pandas==2.2.3
openpyxl==3.1.5

# Document Processing & AI
PyMuPDF==1.26.7
google-generativeai==0.8.5
